requirements.txt
streamlit
openai
google-genai
pypdf
python-dotenv
